flip boot pin default state
I modified the initialization for the boot pin by changing the input resistance from pull-down to pull-up.  This in effect will flip the default state of the device so that it will boot up into the application and not into the bootloader.  To boot up into the bootloader, the user will need to ground the boot pin prior to powering up the device.